requirements.txt
# Community Hospital MLOps Platform - Production Dependencies
# Python 3.9+

# Core Data Science & ML
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.2
xgboost==2.0.3
hyperopt==0.2.7

# Feature Store
feast==0.35.0
protobuf==4.23.4

# Model Registry & Experiment Tracking
mlflow==2.11.3

# Model Explainability
shap==0.45.1

# Data Validation & Quality
pydantic==1.10.7
great-expectations==0.18.8

# Monitoring & Drift Detection
evidently==0.2.8
alibi-detect==0.12.0

# Dashboard & Visualization
streamlit==1.36.0
streamlit-shap==1.0.2
matplotlib==3.7.0
seaborn==0.13.2
plotly==5.18.0

# Data Processing
pyarrow==8.0.0
dask[dataframe]==2023.12.1
pyjanitor==0.27.0

# Big Data (Optional - for Spark integration)
pyspark==3.5.1
delta-spark==3.1.0

# Database
sqlalchemy==2.0.25

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Networking (for MLflow, Feast)
grpcio==1.48.2
grpcio-status==1.48.1

# Jupyter Support
ipython==8.14.0
ipykernel==6.25.0
ipywidgets==8.1.1
jedi==0.18.1
jinja2==3.1.2

# Misc
six==1.16.0
